Refactor devapps package: remove unused files, update build scripts, and enhance configuration handling
