+++ packages/studiocms_devapps/.gitignore
@@ -19,3 +19,5 @@ pnpm-debug.log*
 .DS_Store
 
 .npmrc
+
+dist/
