+++ packages/studiocms_devapps/src/utils/app-utils.ts
