+++ packages/studiocms_devapps/src/virt.d.ts
@@ -1,0 +1,14 @@
+/// <reference types="@astrojs/db" />
+
+declare module 'virtual:studiocms-devapps/endpoints' {
+	export const libSQLEndpoint: string;
+	export const wpAPIEndpoint: string;
+}
+
+declare module 'virtual:studiocms-devapps/config' {
+	export const userProjectRoot: string;
+	export const dbEnv: {
+		remoteUrl: string;
+		token: string;
+	};
+}
